now Create vote and change vote in one API controller
